Fix package names to lowercase
