+++ TicTacToe/src/ai/MiniMaxAlphaBeta.java
@@ -1,29 +1,35 @@
-package AI;
+package ai;
 
-import Game.Board;
+import game.Board;
 
 /**
- * The MiniMax algorithm in its most basic form.
+ * The MiniMax algorithm optimised with Alpha-Beta pruning.
  * @author DavidHurst
  */
-public class MiniMax {
+public class MiniMaxAlphaBeta {
 
-    private static final int MAX_DEPTH = 6;
+    private static final int MAX_DEPTH = 12;
 
-    private MiniMax() {
+    private MiniMaxAlphaBeta() {
     }
 
     /**
      * Play moves on the board alternating between playing as X and O analysing 
      * the board each time to return the value of the highest value move for the
-     * X player. Return the highest value move when a terminal node or the 
-     * maximum search depth is reached.
+     * X player. Use variables alpha and beta as the best alternative for the 
+     * maximising player (X) and the best alternative for the minimising player 
+     * (O) respectively, do not search descendants of nodes if player's 
+     * alternatives are better than the node. Return the highest value move when 
+     * a terminal node or the maximum search depth is reached.
      * @param board Board to play on and evaluate
      * @param depth The maximum depth of the game tree to search to
+     * @param alpha The best alternative for the maximising player (X)
+     * @param beta The best alternative for the minimising player (O)
      * @param isMax Maximising or minimising player 
      * @return Value of the board 
      */
-    public static int miniMax(Board board, int depth, boolean isMax) {
+    public static int miniMax(Board board, int depth, int alpha, int beta,
+            boolean isMax) {
         int boardVal = evaluateBoard(board);
 
         // Terminal node (win/lose/draw) or max depth reached.
@@ -40,8 +46,12 @@ public class MiniMax {
                     if (!board.isTileMarked(row, col)) {
                         board.setMarkAt(row, col, 'X');
                         highestVal = Math.max(highestVal, miniMax(board,
-                                depth - 1, false));
+                                depth - 1, alpha, beta, false));
                         board.setMarkAt(row, col, ' ');
+                        alpha = Math.max(alpha, highestVal);
+                        if (alpha >= beta) {
+                            return highestVal;
+                        }
                     }
                 }
             }
@@ -54,8 +64,12 @@ public class MiniMax {
                     if (!board.isTileMarked(row, col)) {
                         board.setMarkAt(row, col, 'O');
                         lowestVal = Math.min(lowestVal, miniMax(board,
-                                depth - 1, true));
+                                depth - 1, alpha, beta, true));
                         board.setMarkAt(row, col, ' ');
+                        beta = Math.min(beta, lowestVal);
+                        if (beta <= alpha) {
+                            return lowestVal;
+                        }
                     }
                 }
             }
@@ -76,7 +90,8 @@ public class MiniMax {
             for (int col = 0; col < board.getWidth(); col++) {
                 if (!board.isTileMarked(row, col)) {
                     board.setMarkAt(row, col, 'X');
-                    int moveValue = miniMax(board, MAX_DEPTH, false);
+                    int moveValue = miniMax(board, MAX_DEPTH, Integer.MIN_VALUE,
+                            Integer.MAX_VALUE, false);
                     board.setMarkAt(row, col, ' ');
                     if (moveValue > bestValue) {
                         bestMove[0] = row;
@@ -155,5 +170,4 @@ public class MiniMax {
 
         return 0;
     }
-
 }
