+++ TicTacToe/src/game/Board.java
@@ -1,4 +1,4 @@
-package Game;
+package game;
 
 /**
  * @author DavidHurst
